requirements.txt
# numpy, use sudo apt-get install python3-numpy

adafruit-circuitpython-ads1x15
RPi.GPIO
click
paho-mqtt
simple-pid
paramiko

# testing
fake-rpi


# optional: pandas, scipy
